requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.7
django-rest-framework==0.1.0
djangorestframework==3.7.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
pytz==2017.3
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
